Update README.md
### The abstract for the thesis (Aggregated Multiscale Self-Supervised Denoising)

In typical image denoising approaches, both supervised and unsupervised learning methods does not take account of individual image's particular image prior, the noise statistics, or both. The networks learned from external images inherently suffer from a domain gap problem as the image priors and noise statistics can be significantly different from the training and test images. So, it is difficult if the methods primarily requires clean images to train denoising. Furthermore, some images inherently generate significant noise (satellite images of distant galaxies, medical images like MRI images, CT scans, X-Ray images, etc.), and there are no clean images for training. Here the problems dominantly lie with the data delivery system. Our approach takes the noisy images and creates a new version of them with specific pre-processing; by doing so, we make the target pseudo clear image for the deep neural network. We generate multiple versions of these noisy images using interpolation of arrays and train the network to the extent where the network can learn the information on the images without the noises. In practice, the noisy pictures are blurred on three different scales. These blurred versions and the original noisy images are combined together to create a single set. This set captures all the necessary information from all of the four groups. The network architecture uses the concatenation of the module concept to learn the clear images from a versatile perspective. Then we trained the model using the main noisy set as input and the newly created set as the target to predict a much cleaner image from a regular noisy image. This method creates an output image where the structural integrity image is sustained and the noise component is removed from the image.